pyproject.toml
[project]
name = "few-shot-tsum"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "decord>=0.6.0",
    "evaluate>=0.4.6",
    "google-genai>=1.41.0",
    "jinja2>=3.1.6",
    "opencv-python>=4.12.0.88",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.10",
    "pycocoevalcap>=1.2",
    "pydantic-settings>=2.11.0",
    "rouge-score>=0.1.2",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "transformers>=4.56.2",
    "yt-dlp>=2025.9.26",
]

[tool.uv.sources]
src = { path = "src" }
